+++ src/main/java/dandi/dandi/member/web/MemberControllerDocs.java
@@ -3,7 +3,6 @@ package dandi.dandi.member.web;
 import static org.springframework.http.HttpHeaders.AUTHORIZATION;
 
 import dandi.dandi.advice.ExceptionResponse;
-import dandi.dandi.member.application.port.in.MemberBlockCommand;
 import dandi.dandi.member.application.port.in.MemberInfoResponse;
 import dandi.dandi.member.application.port.in.NicknameDuplicationCheckResponse;
 import dandi.dandi.member.web.dto.in.LocationUpdateRequest;
@@ -49,12 +48,4 @@ public interface MemberControllerDocs {
     })
     ResponseEntity<Void> updateMemberLocation(@Parameter(hidden = true) Long memberId,
                                               LocationUpdateRequest locationUpdateRequest);
-
-    @Operation(summary = "사용자 차단")
-    @ApiResponses(value = {
-            @ApiResponse(responseCode = "201", description = "사용자 차단 성공"),
-            @ApiResponse(responseCode = "400", description = "이미 차단한 사용자"),
-            @ApiResponse(responseCode = "404", description = "존재하지 않는 사용자 사용자"),
-    })
-    ResponseEntity<Void> blockMember(@Parameter(hidden = true) Long memberId, MemberBlockCommand memberBlockCommand);
 }
